Revert "Merge branch 'programmer-branch' into PlatformBranch"
This reverts commit 88b3e1016358ce7b4605c7eb67383ffafc958585, reversing
changes made to 22ffe57f1bcc3fa19ec514b0ce943aabd743ab60.
