+++ assets/constants.json
@@ -119,8 +119,15 @@
 			"jazzPos": [0.0, 4.0]
 		}
 	],
+	"mplatforms":   [
+		{
+			"pos": [18.0,2.5,18.5,2.5,18.5,5.5,18.0,5.5],
+			"nodes": [0.0, 0.0, 3.0, 1.0, 3.0, 3.0],
+			"speed": 0.5
+		}
+	],
 	"genre_gravity": {
 		"synth": -14.7,
 		"jazz": -9.0
-	},
+	}
 }
