+++ core/src/edu/cornell/gdiac/rabbeat/WorldController.java
@@ -16,8 +16,16 @@
  */
 package edu.cornell.gdiac.rabbeat;
 
-import edu.cornell.gdiac.rabbeat.obstacles.enemies.SyncedProjectile;
+import edu.cornell.gdiac.rabbeat.obstacle.enemies.BearEnemy;
+
+
+import edu.cornell.gdiac.rabbeat.obstacle.enemies.Enemy;
+
+import edu.cornell.gdiac.rabbeat.obstacle.enemies.SyncedProjectile;
+import edu.cornell.gdiac.rabbeat.obstacle.platforms.MovingPlatform;
+import edu.cornell.gdiac.rabbeat.obstacle.platforms.SyncedPlatform;
 import edu.cornell.gdiac.rabbeat.sync.BeatTest;
+import edu.cornell.gdiac.rabbeat.sync.BulletSync;
 import edu.cornell.gdiac.rabbeat.sync.ISynced;
 import edu.cornell.gdiac.rabbeat.sync.SyncController;
 import java.util.Iterator;
@@ -27,10 +35,13 @@ import com.badlogic.gdx.audio.*;
 import com.badlogic.gdx.math.*;
 import com.badlogic.gdx.utils.*;
 import com.badlogic.gdx.graphics.*;
+import com.badlogic.gdx.graphics.g2d.*;
 import com.badlogic.gdx.physics.box2d.*;
 import edu.cornell.gdiac.assets.AssetDirectory;
+import edu.cornell.gdiac.rabbeat.obstacle.platforms.WeightedPlatform;
 import edu.cornell.gdiac.util.*;
-import edu.cornell.gdiac.rabbeat.obstacles.*;
+import edu.cornell.gdiac.rabbeat.obstacle.*;
+//import javax.xml.soap.Text;
 
 /**
  * Base class for a world-specific controller.
@@ -46,14 +57,25 @@ import edu.cornell.gdiac.rabbeat.obstacles.*;
  * This is the purpose of our AssetState variable; it ensures that multiple instances
  * place nicely with the static assets.
  */
-public class GameController implements Screen, ContactListener {
+public class WorldController implements Screen, ContactListener {
 
 	/** The genre state of the game */
 	public Genre genre = Genre.SYNTH;
 	/** The Sync object that will sync the world to the beat*/
 	public SyncController syncController;
-	/** The object loader for creating objects into the world */
-	public ObjectController objectController;
+
+	/** The texture for walls */
+	protected TextureRegion blackTile;
+	/** The texture for regular platforms */
+	protected TextureRegion platformTile;
+	/** The texture for weighted platforms */
+	protected TextureRegion weightedPlatform;
+	/** The texture for bullets */
+	protected TextureRegion bullet;
+	/** The texture for the exit condition */
+	protected TextureRegion goalTile;
+	/** The font for giving messages to the player */
+	protected BitmapFont displayFont;
 	
 	/** Exit code for quitting the game */
 	public static final int EXIT_QUIT = 0;
@@ -75,11 +97,9 @@ public class GameController implements Screen, ContactListener {
 	/** Reference to the game canvas */
 	protected GameCanvas canvas;
 	/** All the objects in the world. */
-	protected PooledList<GameObject> objects  = new PooledList<>();
-	/** All objects that are genre-dependent*/
-	protected PooledList<IGenreObject> genreObjects = new PooledList<>();
+	protected PooledList<Obstacle> objects  = new PooledList<Obstacle>();
 	/** Queue for adding objects */
-	protected PooledList<GameObject> addQueue = new PooledList<>();
+	protected PooledList<Obstacle> addQueue = new PooledList<Obstacle>();
 	/** Listener that will update the player mode when we are done */
 	private ScreenListener listener;
 
@@ -101,30 +121,67 @@ public class GameController implements Screen, ContactListener {
 	/** Countdown active for winning or losing */
 	private int countdown;
 
+
+	/** Textures for rab-beat*/
+
+	private TextureRegion synthDefaultTexture;
+	private TextureRegion synthJazzTexture;
+	private TextureRegion backgroundTexture;
+	private TextureRegion enemyDefaultTexture;
+
 	// TODO: Add sounds and sound id fields here
 	/**synth soundtrack of game*/
 	private Music synthSoundtrack;
 	/** jazz soundtrack of game*/
 	private Music jazzSoundtrack;
 
+	private float volume;
+
+	/** The player scale for synth */
+	private float playerScale = 3/8f;
+
+	/** The enemy scale for the enemy */
+	private float enemyScale = 3/8f;
 
 	// Physics objects for the game
+	/** Physics constants for initialization */
+	private JsonValue constants;
+	/** Reference to the character avatar */
+	private static Player avatar;
 
-	/** the spawnpoint location of the player*/
+	/** Reference to the enemy avatar */
+	private Enemy enemy;
 
-	private BoxGameObject respawnPoint;
+	/** Reference to the goalDoor (for collision detection) */
+	private BoxObstacle goalDoor;
+	/** Reference to the current spawn point */
+	private BoxObstacle respawnPoint;
+	/** Reference to all the checkpoints */
+	private Queue<Pair<BoxObstacle, Integer>> checkpoints;
+	/** Reference to all the weighted platforms */
+	private Array<SyncedPlatform> weightedPlatforms;
+
+	/** Reference to all the weighted platforms */
+	private Array<Enemy> enemies;
 
 	/** Mark set to handle more sophisticated collision callbacks */
 	protected ObjectSet<Fixture> sensorFixtures;
 
-	private static GameController theController = null;
+	private static WorldController theController = null;
 
-	public static GameController getInstance() {
+	public static WorldController getInstance() {
 		if (theController == null) {
-			theController = new GameController();
+			theController = new WorldController();
 		}
 		return theController;
 	}
+
+	/** Returns the player object */
+	public static Player getPlayer() {
+		return avatar;
+	}
+
+
 
 	/**
 	 * Returns true if debug mode is active.
@@ -239,7 +296,7 @@ public class GameController implements Screen, ContactListener {
 	 * with the Box2d coordinates.  The bounds are in terms of the Box2d
 	 * world, not the screen.
 	 */
-	protected GameController() {
+	protected WorldController() {
 		this(new Rectangle(0,0,DEFAULT_WIDTH,DEFAULT_HEIGHT), 
 			 new Vector2(0,DEFAULT_GRAVITY));
 		setDebug(false);
@@ -247,8 +304,10 @@ public class GameController implements Screen, ContactListener {
 		setFailure(false);
 		world.setContactListener(this);
 		sensorFixtures = new ObjectSet<Fixture>();
+		checkpoints = new Queue<Pair<BoxObstacle, Integer>>();
+		weightedPlatforms = new Array<>();
+		enemies = new Array<>();
 		syncController = new SyncController();
-		objectController = new ObjectController();
 		theController = this;
 	}
 
@@ -262,7 +321,7 @@ public class GameController implements Screen, ContactListener {
 	 * @param bounds	The game bounds in Box2d coordinates
 	 * @param gravity	The gravitational force on this Box2d world
 	 */
-	protected GameController(Rectangle bounds, Vector2 gravity) {
+	protected WorldController(Rectangle bounds, Vector2 gravity) {
 		world = new World(gravity,false);
 		this.bounds = new Rectangle(bounds);
 		this.scale = new Vector2(1,1);
@@ -277,7 +336,7 @@ public class GameController implements Screen, ContactListener {
 	 * Dispose of all (non-static) resources allocated to this mode.
 	 */
 	public void dispose() {
-		for(GameObject obj : objects) {
+		for(Obstacle obj : objects) {
 			obj.deactivatePhysics(world);
 		}
 		objects.clear();
@@ -289,9 +348,6 @@ public class GameController implements Screen, ContactListener {
 		scale  = null;
 		world  = null;
 		canvas = null;
-		syncController = null;
-		genreObjects = null;
-		objectController = null;
 	}
 
 	// TODO: Adjust to the correct assets after assets have been added
@@ -304,9 +360,24 @@ public class GameController implements Screen, ContactListener {
 	 * @param directory	Reference to global asset manager.
 	 */
 	public void gatherAssets(AssetDirectory directory) {
-		objectController.gatherAssets(directory);
+		synthDefaultTexture = new TextureRegion(directory.getEntry("player:synth",Texture.class));
+		synthJazzTexture = new TextureRegion(directory.getEntry("player:synth-jazz",Texture.class));
+
 		//set the soundtrack
 		setSoundtrack(directory);
+
+		backgroundTexture = new TextureRegion(directory.getEntry("backgrounds:test-bg",Texture.class));
+		enemyDefaultTexture = new TextureRegion(directory.getEntry("player:synth",Texture.class)); //CHANGE FOR ENEMY!
+
+		constants = directory.getEntry( "constants", JsonValue.class );
+
+		// Allocate the tiles
+		blackTile = new TextureRegion(directory.getEntry( "world:platforms:blackTile", Texture.class ));
+		platformTile = new TextureRegion(directory.getEntry( "world:platforms:platformTile", Texture.class ));
+		weightedPlatform = new TextureRegion((directory.getEntry("world:platforms:weightedPlatform", Texture.class)));
+		bullet = new TextureRegion(directory.getEntry("world:bullet", Texture.class));
+		goalTile  = new TextureRegion(directory.getEntry( "world:goal", Texture.class ));
+		displayFont = directory.getEntry( "fonts:retro" ,BitmapFont.class);
 	}
 
 	/** Sets the synth and jazz soundtrack to the correct tracks.  This function will be significant
@@ -314,13 +385,21 @@ public class GameController implements Screen, ContactListener {
 	 *
 	 * @param directory	Reference to global asset manager.
 	 * */
-	public void setSoundtrack(AssetDirectory directory) {
-		synthSoundtrack = directory.getEntry("music:synth1", Music.class);
+	public void setSoundtrack(AssetDirectory directory){
+		synthSoundtrack = directory.getEntry("music:synth1", Music.class) ;
 		synthSoundtrack.setLooping(true);
 		synthSoundtrack.setVolume(1);
-		jazzSoundtrack = directory.getEntry("music:jazz1", Music.class);
+		jazzSoundtrack = directory.getEntry("music:jazz1",Music.class);
 		jazzSoundtrack.setLooping(true);
 		jazzSoundtrack.setVolume(0);
+	}
+
+	public JsonValue getBulletJV(){
+		return constants.get("bullet");
+	}
+
+	public TextureRegion getBulletTR(){
+		return bullet;
 	}
 
 	public Vector2 getScale(){
@@ -340,25 +419,20 @@ public class GameController implements Screen, ContactListener {
 	 *
 	 * param obj The object to add
 	 */
-	public void instantiateQueue(GameObject obj) {
+	public void addQueuedObject(Obstacle obj) {
 		assert inBounds(obj) : "Object is not in bounds";
 		addQueue.add(obj);
 	}
-	/**If the object is implements {@link ISynced}, add
-	 * to the sync.  If it is a {@link IGenreObject}, add to genreObstacles.
-	 * @param  object: The object you are instantiating
+
+	/**
+	 * Immediately adds the object to the physics world
 	 *
-	 * */
-	protected void instantiate(GameObject object){
-		assert inBounds(object) : "Object is not in bounds";
-		objects.add(object);
-		if(object instanceof  ISynced){
-			syncController.addSync((ISynced) object);
-		}
-		if(object instanceof IGenreObject){
-			genreObjects.add((IGenreObject) object);
-		}
-		object.activatePhysics(world);
+	 * param obj The object to add
+	 */
+	protected void addObject(Obstacle obj) {
+		assert inBounds(obj) : "Object is not in bounds";
+		objects.add(obj);
+		obj.activatePhysics(world);
 	}
 
 	/**
@@ -370,7 +444,7 @@ public class GameController implements Screen, ContactListener {
 	 *
 	 * @return true if the object is in bounds.
 	 */
-	public boolean inBounds(GameObject obj) {
+	public boolean inBounds(Obstacle obj) {
 		boolean horiz = (bounds.x <= obj.getX() && obj.getX() <= bounds.x+bounds.width);
 		boolean vert  = (bounds.y <= obj.getY() && obj.getY() <= bounds.y+bounds.height);
 		return horiz && vert;
@@ -388,7 +462,49 @@ public class GameController implements Screen, ContactListener {
 		setComplete(false);
 		setFailure(false);
 		populateLevel();
-		objectController.player.setPosition(respawnPoint.getPosition());
+		createCheckpoints();
+		avatar.setPosition(respawnPoint.getPosition());
+	}
+
+	/**
+	 * Create the start tile and checkpoints
+	 */
+	private void createCheckpoints() {
+		float checkpointWidth  = goalTile.getRegionWidth()/scale.x;
+		float checkpointHeight = goalTile.getRegionHeight()/scale.y;
+
+		// Add the start tile as the current spawn point
+		JsonValue start = constants.get("start");
+		JsonValue startPos = start.get("pos");
+		BoxObstacle startTile = new BoxObstacle(startPos.getFloat(0), startPos.getFloat(1), checkpointWidth, checkpointHeight);
+		startTile.setBodyType(BodyDef.BodyType.StaticBody);
+		startTile.setDensity(start.getFloat("density", 0));
+		startTile.setFriction(start.getFloat("friction", 0));
+		startTile.setRestitution(start.getFloat("restitution", 0));
+		startTile.setSensor(true);
+		startTile.setDrawScale(scale);
+		startTile.setTexture(goalTile);
+		startTile.setName("start");
+		addObject(startTile);
+		respawnPoint = startTile;
+
+		// Populate all checkpoints
+		for (int i = 0; i < constants.get("checkpoints").size; i++) {
+			String cname = "checkpoint";
+			JsonValue checkpoint = constants.get("checkpoints").get(i);
+			JsonValue checkpointPos = checkpoint.get("pos");
+			BoxObstacle obj = new BoxObstacle(checkpointPos.getFloat(0), checkpointPos.getFloat(1), checkpointWidth, checkpointHeight);
+			obj.setBodyType(BodyDef.BodyType.StaticBody);
+			obj.setDensity(checkpoint.getFloat("density", 0));
+			obj.setFriction(checkpoint.getFloat("friction", 0));
+			obj.setRestitution(checkpoint.getFloat("restitution", 0));
+			obj.setSensor(true);
+			obj.setDrawScale(scale);
+			obj.setTexture(goalTile);
+			obj.setName(cname + i);
+			addObject(obj);
+			checkpoints.addLast(new Pair<BoxObstacle, Integer>(obj, i));
+		}
 	}
 
 	// TODO: Reset to SYNTH defaults
@@ -402,7 +518,7 @@ public class GameController implements Screen, ContactListener {
 		genre = Genre.SYNTH;
 		Vector2 gravity = new Vector2(world.getGravity() );
 
-		for(GameObject obj : objects) {
+		for(Obstacle obj : objects) {
 			obj.deactivatePhysics(world);
 		}
 		objects.clear();
@@ -415,6 +531,7 @@ public class GameController implements Screen, ContactListener {
 		setFailure(false);
 		syncController = new SyncController();
 		populateLevel();
+		avatar.setPosition(respawnPoint.getPosition());
 	}
 
 	// TODO: Will use level data json to populate
@@ -422,17 +539,143 @@ public class GameController implements Screen, ContactListener {
 	 * Lays out the game geography.
 	 */
 	private void populateLevel() {
+		// Repopulate current checkpoints
+		float checkpointWidth  = goalTile.getRegionWidth()/scale.x;
+		float checkpointHeight = goalTile.getRegionHeight()/scale.y;
+
+		Queue<Pair<BoxObstacle, Integer>> newCheckpoints = new Queue<>();
+		for (Pair<BoxObstacle, Integer> pair : checkpoints) {
+			String cname = "checkpoint";
+			JsonValue checkpoint = constants.get("checkpoints").get(pair.snd);
+			JsonValue checkpointPos = checkpoint.get("pos");
+			BoxObstacle obj = new BoxObstacle(checkpointPos.getFloat(0), checkpointPos.getFloat(1), checkpointWidth, checkpointHeight);
+			obj.setBodyType(BodyDef.BodyType.StaticBody);
+			obj.setDensity(checkpoint.getFloat("density", 0));
+			obj.setFriction(checkpoint.getFloat("friction", 0));
+			obj.setRestitution(checkpoint.getFloat("restitution", 0));
+			obj.setSensor(true);
+			obj.setDrawScale(scale);
+			obj.setTexture(goalTile);
+			obj.setName(cname + pair.snd);
+			addObject(obj);
+			newCheckpoints.addLast(new Pair<BoxObstacle, Integer>(obj, pair.snd));
+		}
+		checkpoints.clear();
+		checkpoints = newCheckpoints;
+
+		// Add level goal
+		float dwidth  = goalTile.getRegionWidth()/scale.x;
+		float dheight = goalTile.getRegionHeight()/scale.y;
+
+		JsonValue goal = constants.get("goal");
+		JsonValue goalpos = goal.get("pos");
+		goalDoor = new BoxObstacle(goalpos.getFloat(0),goalpos.getFloat(1),dwidth,dheight);
+		goalDoor.setBodyType(BodyDef.BodyType.StaticBody);
+		goalDoor.setDensity(goal.getFloat("density", 0));
+		goalDoor.setFriction(goal.getFloat("friction", 0));
+		goalDoor.setRestitution(goal.getFloat("restitution", 0));
+		goalDoor.setSensor(true);
+		goalDoor.setDrawScale(scale);
+		goalDoor.setTexture(goalTile);
+		goalDoor.setName("goal");
+		instantiate(goalDoor);
+
+		String wname = "wall";
+		JsonValue walljv = constants.get("walls");
+		JsonValue defaults = constants.get("defaults");
+		for (int ii = 0; ii < walljv.size; ii++) {
+			PolygonObstacle obj;
+			obj = new PolygonObstacle(walljv.get(ii).asFloatArray(), 0, 0);
+			obj.setBodyType(BodyDef.BodyType.StaticBody);
+			obj.setDensity(defaults.getFloat( "density", 0.0f ));
+			obj.setFriction(defaults.getFloat( "friction", 0.0f ));
+			obj.setRestitution(defaults.getFloat( "restitution", 0.0f ));
+			obj.setDrawScale(scale);
+			obj.setTexture(blackTile);
+			obj.setName(wname+ii);
+			instantiate(obj);
+		}
+
+		String pname = "platform";
+		JsonValue platjv = constants.get("platforms");
+		for (int ii = 0; ii < platjv.size; ii++) {
+			PolygonObstacle obj;
+			obj = new PolygonObstacle(platjv.get(ii).asFloatArray(), 0, 0);
+			obj.setBodyType(BodyDef.BodyType.StaticBody);
+			obj.setDensity(defaults.getFloat( "density", 0.0f ));
+			obj.setFriction(defaults.getFloat( "friction", 0.0f ));
+			obj.setRestitution(defaults.getFloat( "restitution", 0.0f ));
+			obj.setDrawScale(scale);
+			obj.setTexture(platformTile);
+			obj.setName(pname+ii);
+			instantiate(obj);
+		}
+
+		String wpname = "wplatform";
+		JsonValue wplatjv = constants.get("wplatforms");
+		for (int ii = 0; ii < wplatjv.size; ii++) {
+			JsonValue currentWP = wplatjv.get(ii);
+			WeightedPlatform obj;
+			obj = new WeightedPlatform(currentWP.get("pos").asFloatArray(), currentWP.get("synthPos").asFloatArray(),
+					currentWP.get("jazzPos").asFloatArray(), 5.0f);
+			obj.setBodyType(BodyDef.BodyType.StaticBody);
+			obj.setDensity(defaults.getFloat("density", 0.0f));
+			obj.setFriction(defaults.getFloat("friction", 0.0f));
+			obj.setRestitution(defaults.getFloat("restitution", 0.0f));
+			obj.setDrawScale(scale);
+			obj.setTexture(weightedPlatform);
+			obj.setName(wpname + ii);
+			instantiate(obj);
+			weightedPlatforms.add(obj);
+		}
+
+		String mpname = "mplatform";
+		JsonValue mplatjv = constants.get("mplatforms");
+		for (int ii = 0; ii < mplatjv.size; ii++) {
+			JsonValue currentWP = mplatjv.get(ii);
+			MovingPlatform obj;
+			obj = new MovingPlatform(currentWP.get("pos").asFloatArray(), currentWP.get("nodes").asFloatArray(),
+					currentWP.getFloat("speed"));
+			obj.setBodyType(BodyDef.BodyType.StaticBody);
+			obj.setDensity(defaults.getFloat("density", 0.0f));
+			obj.setFriction(defaults.getFloat("friction", 0.0f));
+			obj.setRestitution(defaults.getFloat("restitution", 0.0f));
+			obj.setDrawScale(scale);
+			obj.setTexture(weightedPlatform);
+			obj.setName(mpname + ii);
+			instantiate(obj);
+			weightedPlatforms.add(obj);
+		}
 
 		//world starts with Synth gravity
-		world.setGravity( new Vector2(0,objectController.constants.get("genre_gravity").getFloat("synth",0)) );
-		//TODO This volume constant is never used
-		float volume = objectController.constants.getFloat("volume", 1.0f);
+		world.setGravity( new Vector2(0,constants.get("genre_gravity").getFloat("synth",0)) );
+
+		// Create bunny
+		dwidth  = synthDefaultTexture.getRegionWidth()/scale.x;
+		dheight = synthDefaultTexture.getRegionHeight()/scale.y;
+		avatar = new Player(constants.get("bunny"), dwidth*playerScale, dheight*playerScale, playerScale);
+		avatar.setDrawScale(scale);
+		avatar.setTexture(synthDefaultTexture);
+		instantiate(avatar);
+
+
+		//TODO: Load enemies
+		dwidth  = enemyDefaultTexture.getRegionWidth()/scale.x;
+		dheight = enemyDefaultTexture.getRegionHeight()/scale.y;
+
+		enemy = new BearEnemy(constants.get("enemy"), dwidth*enemyScale, dheight*enemyScale, enemyScale, false,
+				constants.get("bullet").getFloat("synth speed", 0), constants.get("bullet").getFloat("jazz speed", 0));
+		enemy.setDrawScale(scale);
+		enemy.setTexture(enemyDefaultTexture);
+		enemies.add(enemy);
+		instantiate(enemy);
+
+		volume = constants.getFloat("volume", 1.0f);
 
 		syncController.addSync(new BeatTest());
 		syncController.setSync(synthSoundtrack, jazzSoundtrack);
 		//TODO: soundtrack play should be controller by soundController
 		synthSoundtrack.play();
-		objectController.populateObjects(scale);;
 
 	}
 
@@ -480,7 +723,7 @@ public class GameController implements Screen, ContactListener {
 				}
 			}
 		}
-		if (!isFailure() && objectController.player.getY() < -1) {
+		if (!isFailure() && avatar.getY() < -1) {
 			setFailure(true);
 			return false;
 		}
@@ -499,15 +742,24 @@ public class GameController implements Screen, ContactListener {
 	 * @param dt	Number of seconds since last animation frame
 	 */
 	public void update(float dt) {
-		//TODO: bullet stuff needs to go and make the update in the object itself, not here
-
+		// Process actions in object model
+		avatar.setMovement(InputController.getInstance().getHorizontal() * avatar.getForce());
+		avatar.setJumping(InputController.getInstance().didPrimary());
+		avatar.applyForce();
 		if (InputController.getInstance().getSwitchGenre()) {
 			switchGenre();
 			InputController.getInstance().setSwitchGenre(false);
 			updateGenreSwitch();
 		}
 		syncController.updateBeat();
+		enemy.switchState(); //when more enemies will be added, this will be in a for-loop
 	}
+
+	/** Removes bullet */
+	public void removeBullet(SyncedProjectile bullet) {
+		bullet.markRemoved(true);
+	}
+
 	/**
 	 * Callback method for the start of a collision
 	 *
@@ -529,29 +781,50 @@ public class GameController implements Screen, ContactListener {
 		Object fd2 = fix2.getUserData();
 
 		try {
-			GameObject bd1 = (GameObject)body1.getUserData();
-			GameObject bd2 = (GameObject)body2.getUserData();
+			Obstacle bd1 = (Obstacle)body1.getUserData();
+			Obstacle bd2 = (Obstacle)body2.getUserData();
 
-			if ((objectController.player.getSensorName().equals(fd2) && objectController.player != bd1) ||
-					(objectController.player.getSensorName().equals(fd1) && objectController.player != bd2)) {
-				objectController.player.setGrounded(true);
-				sensorFixtures.add(objectController.player == bd1 ? fix2 : fix1); // Could have more than one ground
+			// Test bullet collision with world
+			if (bd1.getName().equals("bullet") && bd2 != enemy) {
+				removeBullet((SyncedProjectile) bd1);
 			}
+
+			if (bd2.getName().equals("bullet") && bd1 != enemy) {
+				removeBullet((SyncedProjectile) bd2);
+			}
+
+//			if (bd1.getName().equals("bullet") && bd2 == avatar){
+//				reset();
+//			}
+//
+//			if (bd2.getName().equals("bullet") && bd1 == avatar){
+//				reset();
+//			}
+
+			// See if we have landed on the ground.
+			if ((avatar.getSensorName().equals(fd2) && avatar != bd1) ||
+					(avatar.getSensorName().equals(fd1) && avatar != bd2)) {
+				avatar.setGrounded(true);
+				sensorFixtures.add(avatar == bd1 ? fix2 : fix1); // Could have more than one ground
+			}
+
 			// Check for win condition
-			if ((bd1 == objectController.player && bd2 == objectController.goalDoor) ||
-					(bd1 == objectController.goalDoor && bd2 == objectController.player)) {
+			if ((bd1 == avatar && bd2 == goalDoor) ||
+					(bd1 == goalDoor && bd2 == avatar)) {
 				setComplete(true);
 			}
 
-			if ((bd1.equals(objectController.player)   && bd2.equals(objectController.enemy))) {
+
+			if ((bd1 == avatar   && bd2 == enemy)) {
 				setFailure(true);
 			}
 
 			// Check for collision with checkpoints and set new current checkpoint
-			if (!objectController.checkpoints.isEmpty() &&
-					((bd1 == objectController.player && bd2 == objectController.checkpoints.first().fst) ||
-					(bd1 == objectController.checkpoints.first().fst && bd2 == objectController.player))) {
-				respawnPoint = objectController.checkpoints.removeFirst().fst;
+			if (!checkpoints.isEmpty() &&
+					((bd1 == avatar && bd2 == checkpoints.first().fst) ||
+					(bd1 == checkpoints.first().fst && bd2 == avatar))) {
+				respawnPoint = checkpoints.removeFirst().fst;
+
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -579,11 +852,11 @@ public class GameController implements Screen, ContactListener {
 		Object bd1 = body1.getUserData();
 		Object bd2 = body2.getUserData();
 
-		if ((objectController.player.getSensorName().equals(fd2) && objectController.player != bd1) ||
-				(objectController.player.getSensorName().equals(fd1) && objectController.player != bd2)) {
-			sensorFixtures.remove(objectController.player == bd1 ? fix2 : fix1);
+		if ((avatar.getSensorName().equals(fd2) && avatar != bd1) ||
+				(avatar.getSensorName().equals(fd1) && avatar != bd2)) {
+			sensorFixtures.remove(avatar == bd1 ? fix2 : fix1);
 			if (sensorFixtures.size == 0) {
-				objectController.player.setGrounded(false);
+				avatar.setGrounded(false);
 			}
 		}
 	}
@@ -601,19 +874,27 @@ public class GameController implements Screen, ContactListener {
 	public void updateGenreSwitch() {
 		//update to Synth
 		if (genre == Genre.SYNTH) {
-			world.setGravity( new Vector2(0,objectController.constants.get("genre_gravity").getFloat("synth",0)) );
+			world.setGravity( new Vector2(0,constants.get("genre_gravity").getFloat("synth",0)) );
+			avatar.setMaxSpeed(constants.get("bunny").get("max_speed").getFloat("synth"));
+			avatar.setTexture(synthDefaultTexture);
+			for (SyncedPlatform platform : weightedPlatforms) {
+				platform.genreUpdate(Genre.SYNTH);
+			}
+			for (Enemy e : enemies) {
+				e.genreUpdate(Genre.SYNTH);
+			}
 		}
 		//update to Jazz
 		else {
-			world.setGravity( new Vector2(0,objectController.constants.get("genre_gravity").getFloat("jazz",0)) );
-		}
-
-		for (IGenreObject g : genreObjects) {
-			g.genreUpdate(genre);
-		}
-		//TODO: Make the bullets inherit IGenreObject so we don't do the double genre check
-		for (SyncedProjectile projectile : objectController.bullets) {
-			projectile.genreUpdate(genre);
+			world.setGravity( new Vector2(0,constants.get("genre_gravity").getFloat("jazz",0)) );
+			avatar.setMaxSpeed(constants.get("bunny").get("max_speed").getFloat("jazz"));
+			avatar.setTexture(synthJazzTexture);
+			for (SyncedPlatform platform : weightedPlatforms) {
+				platform.genreUpdate(Genre.JAZZ);
+			}
+			for (Enemy e : enemies) {
+				e.genreUpdate(Genre.JAZZ);
+			}
 		}
 	}
 
@@ -638,10 +919,10 @@ public class GameController implements Screen, ContactListener {
 		// Garbage collect the deleted objects.
 		// Note how we use the linked list nodes to delete O(1) in place.
 		// This is O(n) without copying.
-		Iterator<PooledList<GameObject>.Entry> iterator = objects.entryIterator();
+		Iterator<PooledList<Obstacle>.Entry> iterator = objects.entryIterator();
 		while (iterator.hasNext()) {
-			PooledList<GameObject>.Entry entry = iterator.next();
-			GameObject obj = entry.getValue();
+			PooledList<Obstacle>.Entry entry = iterator.next();
+			Obstacle obj = entry.getValue();
 			if (obj.isRemoved()) {
 				obj.deactivatePhysics(world);
 				entry.remove();
@@ -667,23 +948,23 @@ public class GameController implements Screen, ContactListener {
 
 		// Draw background unscaled.
 		canvas.begin();
-		canvas.draw(objectController.backgroundTexture, Color.WHITE, 0, 0,canvas.getWidth(),canvas.getHeight());
+		canvas.draw(backgroundTexture, Color.WHITE, 0, 0,canvas.getWidth(),canvas.getHeight());
 		canvas.end();
 		
 		canvas.begin();
-		for (GameObject obj : objects) {
+		for (Obstacle obj : objects) {
 			obj.draw(canvas);
 		}
 		canvas.end();
 
 		// Draw the player on top
 		canvas.begin();
-		objectController.player.draw(canvas);
+		avatar.draw(canvas);
 		canvas.end();
 		
 		if (debug) {
 			canvas.beginDebug();
-			for(GameObject obj : objects) {
+			for(Obstacle obj : objects) {
 				obj.drawDebug(canvas);
 			}
 			canvas.endDebug();
@@ -691,14 +972,14 @@ public class GameController implements Screen, ContactListener {
 		
 		// Final message
 		if (complete && !failed) {
-			objectController.displayFont.setColor(Color.YELLOW);
+			displayFont.setColor(Color.YELLOW);
 			canvas.begin(); // DO NOT SCALE
-			canvas.drawTextCentered("VICTORY!", objectController.displayFont, 0.0f);
+			canvas.drawTextCentered("VICTORY!", displayFont, 0.0f);
 			canvas.end();
 		} else if (failed) {
-			objectController.displayFont.setColor(Color.RED);
+			displayFont.setColor(Color.RED);
 			canvas.begin(); // DO NOT SCALE
-			canvas.drawTextCentered("FAILURE!", objectController.displayFont, 0.0f);
+			canvas.drawTextCentered("FAILURE!", displayFont, 0.0f);
 			canvas.end();
 		}
 	}
@@ -771,7 +1052,7 @@ public class GameController implements Screen, ContactListener {
 				update(delta); // This is the one that must be defined.
 				postUpdate(delta);
 			}
-			canvas.updateCamera(objectController.player);
+			canvas.updateCamera(avatar);
 			draw(delta);
 		}
 	}
@@ -837,12 +1118,17 @@ public class GameController implements Screen, ContactListener {
 		}
 	}
 
-	public void setSpawn(BoxGameObject spawn){
-		respawnPoint = spawn;
-	}
+	/**Instantiate an object into the world.  If the object is implements {@link ISynced}, add
+	 * to the sync
+	 * @param  object: The object you are instantiating
+	 *
+	 * */
+	public void instantiate(Obstacle object){
+		addObject(object);
+		if(object instanceof  ISynced){
+			syncController.addSync((ISynced) object);
+		}
 
-	public Player getPlayer(){
-		return objectController.player;
 	}
