+++ core/src/edu/cornell/gdiac/rabbeat/obstacle/Obstacle.java
@@ -1,11 +1,12 @@
 /*
- * SimpleObstacle.java
+ * Obstacle.java
  *
- * This class is a subclass of Obstacle that supports only one Body.
- * it is the prime subclass of most models in the game.
+ * This class and all of the others in this package are provided to give you easy 
+ * access to Box2d.  Box2d's separation of bodies and shapes (e.g. fixtures) can 
+ * be a bit daunting. This class combines them together to simplify matters.
  *
- * This class does not provide Shape information, and cannot be instantiated
- * directly.
+ * This, and its superclasses, are fairly robust.  You may want to use
+ * them in your game.  They make Box2D a lot easier.
  *
  * Many of the method comments in this class are taken from the Box2d manual by 
  * Erin Catto (2011).
@@ -14,34 +15,59 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * LibGDX version, 2/6/2015
  */
-package edu.cornell.gdiac.rabbeat.obstacles;
+package edu.cornell.gdiac.rabbeat.obstacle;
 
 import com.badlogic.gdx.math.*;
-import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.g2d.*;
 import com.badlogic.gdx.physics.box2d.*;
 import com.badlogic.gdx.physics.box2d.BodyDef.BodyType;
 
-import edu.cornell.gdiac.rabbeat.GameCanvas;
+import edu.cornell.gdiac.rabbeat.*;  // For GameCanvas
 
 /**
  * Base model class to support collisions.
  *
- * This is an instance of a Obstacle with just one body. It does not have any joints.
- * It is the primary type of physics object. 
+ * Instances represents a body and/or a group of bodies.
+ * There should be NO game controlling logic code in a physics objects, 
+ * that should reside in the Controllers.
  *
- * This class does not provide Shape information, and cannot be instantiated directly.
+ * This abstract class has no Body or Shape information and should never 
+ * be instantiated directly. Instead, you should instantiate either
+ * SimplePhysicsObject or ComplexPhysicsObject.  This class only exists
+ * to unify common functionality. In particular, it wraps the body and 
+ * and fixture information into a single interface.
  */
-public abstract class SimpleGameObject extends GameObject {
-	/** The physics body for Box2D. */
-	protected Body body;
+public abstract class Obstacle {
+	/// Initialization structures to store body information
+	/** Stores the body information for this shape */
+	protected BodyDef bodyinfo;
+	/** Stores the fixture information for this shape */
+	protected FixtureDef fixture;
+	/** The mass data of this shape (which may override the fixture) */
+	protected MassData massdata;
+	/** Whether or not to use the custom mass data */
+	protected boolean masseffect;
+    /** A tag for debugging purposes */
+    private String nametag;
+	/** Drawing scale to convert physics units to pixels */
+	protected Vector2 drawScale;
 
-	/** The texture for the shape. */
-	protected TextureRegion texture;
-
-	/** The texture origin for drawing */
-	protected Vector2 origin;
+	/// Track garbage collection status
+	/** Whether the object should be removed from the world on next pass */
+	private boolean toRemove;
+	/** Whether the object has changed shape and needs a new fixture */
+	private boolean isDirty;
 	
+	/// Caching objects
+	/** A cache value for when the user wants to access the body position */
+	protected Vector2 positionCache = new Vector2();
+	/** A cache value for when the user wants to access the linear velocity */
+	protected Vector2 velocityCache = new Vector2();
+	/** A cache value for when the user wants to access the center of mass */
+	protected Vector2 centroidCache = new Vector2();
+	/** A cache value for when the user wants to access the drawing scale */
+	protected Vector2 scaleCache = new Vector2();
+
+
 	/// BodyDef Methods
 	/**
 	 * Returns the body type for Box2D physics
@@ -53,7 +79,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the body type for Box2D physics
 	 */
 	public BodyType getBodyType() {
-		return (body != null ? body.getType() : super.getBodyType());
+		return bodyinfo.type;
 	}
 	
 	/**
@@ -66,11 +92,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the body type for Box2D physics
 	 */
 	public void setBodyType(BodyType value) {
-		if (body != null) {
-			body.setType(value);
-		} else {
-			super.setBodyType(value);
-		}
+		bodyinfo.type = value;
 	}
 	
 	/**
@@ -83,7 +105,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the current position for this physics body
 	 */
 	public Vector2 getPosition() {
-		return (body != null ? body.getPosition() : super.getPosition());
+		return positionCache.set(bodyinfo.position);
 	}
 	
 	/**
@@ -94,11 +116,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the current position for this physics body
 	 */
 	public void setPosition(Vector2 value) {
-		if (body != null) {
-			body.setTransform(value,body.getAngle());
-		} else {
-			super.setPosition(value);
-		}
+		bodyinfo.position.set(value);
 	}
 
 	/**
@@ -108,12 +126,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param y  the y-coordinate for this physics body
 	 */
 	public void setPosition(float x, float y) {
-		if (body != null) {
-			positionCache.set(x,y);
-			body.setTransform(positionCache,body.getAngle());
-		} else {
-			super.setPosition(x,y);
-		}
+		bodyinfo.position.set(x, y);
 	}
 
 	/**
@@ -122,7 +135,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the x-coordinate for this physics body
 	 */
 	public float getX() {
-		return (body != null ? body.getPosition().x : super.getX());
+		return bodyinfo.position.x;
 	}
 	
 	/**
@@ -131,12 +144,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the x-coordinate for this physics body
 	 */
 	public void setX(float value) {
-		if (body != null) {
-			positionCache.set(value,body.getPosition().y);
-			body.setTransform(positionCache,body.getAngle());
-		} else {
-			super.setX(value);
-		}
+		bodyinfo.position.x = value;
 	}
 	
 	/**
@@ -145,7 +153,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the y-coordinate for this physics body
 	 */
 	public float getY() {
-		return (body != null ? body.getPosition().y : super.getY());
+		return bodyinfo.position.y;
 	}
 	
 	/**
@@ -154,12 +162,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the y-coordinate for this physics body
 	 */
 	public void setY(float value) {
-		if (body != null) {
-			positionCache.set(body.getPosition().x,value);
-			body.setTransform(positionCache,body.getAngle());
-		} else {
-			super.setY(value);
-		}
+		bodyinfo.position.y = value;
 	}
 	
 	/**
@@ -170,7 +173,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the angle of rotation for this body
 	 */
 	public float getAngle() {
-		return (body != null ? body.getAngle() : super.getAngle());
+		return bodyinfo.angle;
 	}
 	
 	/**
@@ -179,11 +182,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the angle of rotation for this body (in radians)
 	 */
 	public void setAngle(float value) {
-		if (body != null) {
-			body.setTransform(body.getPosition(),value);
-		} else {
-			super.setAngle(value);
-		}
+		bodyinfo.angle = value;
 	}
 	
 	/**
@@ -196,7 +195,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the linear velocity for this physics body
 	 */
 	public Vector2 getLinearVelocity() {
-		return (body != null ? body.getLinearVelocity() : super.getLinearVelocity());
+		return velocityCache.set(bodyinfo.linearVelocity);
 	}
 	
 	/**
@@ -207,11 +206,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the linear velocity for this physics body
 	 */
 	public void setLinearVelocity(Vector2 value) {
-		if (body != null) {
-			body.setLinearVelocity(value);
-		} else {
-			super.setLinearVelocity(value);
-		}
+		bodyinfo.linearVelocity.set(value);
 	}
 	
 	/**
@@ -220,7 +215,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the x-velocity for this physics body
 	 */
 	public float getVX() {
-		return (body != null ? body.getLinearVelocity().x : super.getVX());
+		return bodyinfo.linearVelocity.x;
 	}
 	
 	/**
@@ -229,12 +224,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the x-velocity for this physics body
 	 */
 	public void setVX(float value) {
-		if (body != null) {
-			velocityCache.set(value,body.getLinearVelocity().y);
-			body.setLinearVelocity(velocityCache);
-		} else {
-			super.setVX(value);
-		}
+		bodyinfo.linearVelocity.x = value;
 	}
 	
 	/**
@@ -243,7 +233,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the y-velocity for this physics body
 	 */
 	public float getVY() {
-		return (body != null ? body.getLinearVelocity().y : super.getVY());
+		return bodyinfo.linearVelocity.y;
 	}
 	
 	/**
@@ -252,12 +242,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the y-velocity for this physics body
 	 */
 	public void setVY(float value) {
-		if (body != null) {
-			velocityCache.set(body.getLinearVelocity().x,value);
-			body.setLinearVelocity(velocityCache);
-		} else {
-			super.setVY(value);
-		}
+		bodyinfo.linearVelocity.y = value;
 	}
 	
 	/**
@@ -268,7 +253,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the angular velocity for this physics body
 	 */
 	public float getAngularVelocity() {
-		return (body != null ? body.getAngularVelocity() : super.getAngularVelocity());
+		return bodyinfo.angularVelocity;
 	}
 	
 	/**
@@ -277,11 +262,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value the angular velocity for this physics body (in radians)
 	 */
 	public void setAngularVelocity(float value) {
-		if (body != null) {
-			body.setAngularVelocity(value);
-		} else {
-			super.setAngularVelocity(value);
-		}
+		bodyinfo.angularVelocity = value;
 	}
 	
 	/**
@@ -295,7 +276,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return true if the body is active
 	 */
 	public boolean isActive() {
-		return (body != null ? body.isActive() : super.isActive());
+		return bodyinfo.active;
 	}
 	
 	/**
@@ -309,11 +290,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  whether the body is active
 	 */
 	public void setActive(boolean value) {
-		if (body != null) {
-			body.setActive(value);
-		} else {
-			super.setActive(value);
-		}
+		bodyinfo.active = value;
 	}
 	
 	/**
@@ -327,7 +304,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return true if the body is awake
 	 */
 	public boolean isAwake() {
-		return (body != null ? body.isAwake() : super.isAwake());
+		return bodyinfo.awake;
 	}
 	
 	/**
@@ -341,11 +318,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  whether the body is awake
 	 */
 	public void setAwake(boolean value) {
-		if (body != null) {
-			body.setAwake(value);
-		} else {
-			super.setAwake(value);
-		}
+		bodyinfo.awake = value;
 	}
 	
 	/**
@@ -359,7 +332,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return false if this body should never fall asleep
 	 */
 	public boolean isSleepingAllowed() {
-		return (body != null ? body.isSleepingAllowed() : super.isSleepingAllowed());
+		return bodyinfo.allowSleep;
 	}
 	
 	/**
@@ -373,11 +346,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  whether the body should ever fall asleep
 	 */
 	public void setSleepingAllowed(boolean value) {
-		if (body != null) {
-			body.setSleepingAllowed(value);
-		} else {
-			super.setSleepingAllowed(value);
-		}
+		bodyinfo.allowSleep = value;
 	}
 	
 	/**
@@ -388,7 +357,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return true if this body be prevented from rotating
 	 */
 	public boolean isFixedRotation() {
-		return (body != null ? body.isFixedRotation() : super.isFixedRotation());
+		return bodyinfo.fixedRotation;
 	}
 	
 	/**
@@ -399,11 +368,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  whether this body be prevented from rotating
 	 */
 	public void setFixedRotation(boolean value) {
-		if (body != null) {
-			body.setFixedRotation(value);
-		} else {
-			super.setFixedRotation(value);
-		}
+		bodyinfo.fixedRotation = value;
 	}
 	
 	/**
@@ -415,7 +380,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the gravity scale to apply to this body
 	 */
 	public float getGravityScale() {
-		return (body != null ? body.getGravityScale() : super.getGravityScale());
+		return bodyinfo.gravityScale;
 	}
 	
 	/**
@@ -427,11 +392,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the gravity scale to apply to this body
 	 */
 	public void setGravityScale(float value) {
-		if (body != null) {
-			body.setGravityScale(value);
-		} else {
-			super.setGravityScale(value);
-		}
+		bodyinfo.gravityScale = value;
 	}
 	
 	/** 
@@ -448,7 +409,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the linear damping for this body.
 	 */
 	public float getLinearDamping() {
-		return (body != null ? body.getLinearDamping() : super.getLinearDamping());
+		return bodyinfo.linearDamping;
 	}
 	
 	/** 
@@ -465,11 +426,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the linear damping for this body.
 	 */
 	public void setLinearDamping(float value) {
-		if (body != null) {
-			body.setLinearDamping(value);
-		} else {
-			super.setLinearDamping(value);
-		}
+		bodyinfo.linearDamping = value;
 	}
 	
 	/** 
@@ -486,7 +443,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the angular damping for this body.
 	 */
 	public float getAngularDamping() {
-		return (body != null ? body.getAngularDamping() : super.getAngularDamping());
+		return bodyinfo.angularDamping;
 	}
 
 	/** 
@@ -503,14 +460,43 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the angular damping for this body.
 	 */
 	public void setAngularDamping(float value) {
-		if (body != null) {
-			body.setAngularDamping(value);
-		} else {
-			super.setAngularDamping(value);
-		}
+		bodyinfo.angularDamping = value;
+	}
+	
+	/**
+	 * Copies the state from the given body to the body def.
+	 *
+	 * This is important if you want to save the state of the body before removing
+	 * it from the world.
+	 */
+	protected void setBodyState(Body body) {
+		bodyinfo.type   = body.getType();
+		bodyinfo.angle  = body.getAngle();
+		bodyinfo.active = body.isActive();
+		bodyinfo.awake  = body.isAwake();
+		bodyinfo.position.set(body.getPosition());
+		bodyinfo.linearVelocity.set(body.getLinearVelocity());
+		bodyinfo.allowSleep = body.isSleepingAllowed();
+		bodyinfo.fixedRotation = body.isFixedRotation();
+		bodyinfo.gravityScale  = body.getGravityScale();
+		bodyinfo.angularDamping = body.getAngularDamping();
+		bodyinfo.linearDamping  = body.getLinearDamping();
 	}
 	
 	/// FixtureDef Methods
+	/**
+	 * Returns the density of this body
+	 *
+	 * The density is typically measured in usually in kg/m^2. The density can be zero or 
+	 * positive. You should generally use similar densities for all your fixtures. This 
+	 * will improve stacking stability.
+	 *
+	 * @return the density of this body
+	 */
+	public float getDensity() {
+		return fixture.density;
+	}
+	
 	/**
 	 * Sets the density of this body
 	 *
@@ -521,12 +507,22 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the density of this body
 	 */
 	public void setDensity(float value) {
-		super.setDensity(value);
-		if (body != null) {
-			for(Fixture f : body.getFixtureList()) {
-				f.setDensity(value);
-			}
-		}
+		fixture.density = value;
+	}
+	
+	/**
+	 * Returns the friction coefficient of this body
+	 *
+	 * The friction parameter is usually set between 0 and 1, but can be any non-negative 
+	 * value. A friction value of 0 turns off friction and a value of 1 makes the friction 
+	 * strong. When the friction force is computed between two shapes, Box2D must combine 
+	 * the friction parameters of the two parent fixtures. This is done with the geometric 
+	 * mean.
+	 *
+	 * @return the friction coefficient of this body
+	 */
+	public float getFriction() {
+		return fixture.friction; 
 	}
 	
 	/**
@@ -541,12 +537,22 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the friction coefficient of this body
 	 */
 	public void setFriction(float value) {
-		super.setFriction(value);
-		if (body != null) {
-			for(Fixture f : body.getFixtureList()) {
-				f.setFriction(value);
-			}
-		}
+		fixture.friction = value;
+	}
+	
+	/**
+	 * Returns the restitution of this body
+	 *
+	 * Restitution is used to make objects bounce. The restitution value is usually set 
+	 * to be between 0 and 1. Consider dropping a ball on a table. A value of zero means 
+	 * the ball won't bounce. This is called an inelastic collision. A value of one means 
+	 * the ball's velocity will be exactly reflected. This is called a perfectly elastic 
+	 * collision.
+	 *
+	 * @return the restitution of this body
+	 */
+	public float getRestitution() {
+		return fixture.restitution;
 	}
 	
 	/**
@@ -561,12 +567,20 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the restitution of this body
 	 */
 	public void setRestitution(float value) {
-		super.setRestitution(value);
-		if (body != null) {
-			for(Fixture f : body.getFixtureList()) {
-				f.setRestitution(value);
-			}
-		}
+		fixture.restitution = value;
+	}
+	
+	/**
+	 * Returns true if this object is a sensor.
+	 *
+	 * Sometimes game logic needs to know when two entities overlap yet there should be 
+	 * no collision response. This is done by using sensors. A sensor is an entity that 
+	 * detects collision but does not produce a response.
+	 *
+	 * @return true if this object is a sensor.
+	 */
+	public boolean isSensor() {
+		return fixture.isSensor;
 	}
 	
 	/**
@@ -579,12 +593,22 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  whether this object is a sensor.
 	 */
 	public void setSensor(boolean value) {
-		super.setSensor(value);
-		if (body != null) {
-			for(Fixture f : body.getFixtureList()) {
-				f.setSensor(value);
-			}
-		}
+		fixture.isSensor = value;
+	}
+	
+	/**
+	 * Returns the filter data for this object (or null if there is none)
+	 *
+	 * Collision filtering allows you to prevent collision between fixtures. For example, 
+	 * say you make a character that rides a bicycle. You want the bicycle to collide 
+	 * with the terrain and the character to collide with the terrain, but you don't want 
+	 * the character to collide with the bicycle (because they must overlap). Box2D 
+	 * supports such collision filtering using categories and groups.
+	 * 
+	 * @return the filter data for this object (or null if there is none)
+	 */
+	public Filter getFilterData() {
+		return fixture.filter;
 	}
 	
 	/**
@@ -601,11 +625,14 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the filter data for this object
 	 */
 	public void setFilterData(Filter value) {
-		super.setFilterData(value);
-		if (body != null) {
-			for(Fixture f : body.getFixtureList()) {
-				f.setFilterData(value);
-			}
+		if (value !=  null) {
+			fixture.filter.categoryBits = value.categoryBits;
+			fixture.filter.groupIndex = value.groupIndex;
+			fixture.filter.maskBits   = value.maskBits;
+		} else {
+			fixture.filter.categoryBits = 0x0001;
+			fixture.filter.groupIndex = 0;
+			fixture.filter.maskBits   = -1;
 		}
 	}
 
@@ -620,7 +647,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the center of mass for this physics body
 	 */
 	public Vector2 getCentroid() {
-		return  (body != null ? body.getLocalCenter() : super.getCentroid());
+		return centroidCache.set(massdata.center);
 	}
 	
 	/**
@@ -631,10 +658,12 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the center of mass for this physics body
 	 */
 	public void setCentroid(Vector2 value) {
-		super.setCentroid(value);
-		if (body != null) {
-			body.setMassData(massdata); // Protected accessor?
+		if (!masseffect) {
+			masseffect = true;
+			massdata.I = getInertia();
+			massdata.mass = getMass();
 		}
+		massdata.center.set(value);
 	}
 	
 	/**
@@ -646,7 +675,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the rotational inertia of this body
 	 */
 	public float getInertia() {
-		return  (body != null ? body.getInertia() : super.getInertia());
+		return massdata.I;
 	}
 
 	/**
@@ -658,10 +687,12 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the rotational inertia of this body
 	 */
 	public void setInertia(float value) {
-		super.setInertia(value);
-		if (body != null) {
-			body.setMassData(massdata); // Protected accessor?
+		if (!masseffect) {
+			masseffect = true;
+			massdata.center.set(getCentroid());
+			massdata.mass = getMass();
 		}
+		massdata.I = value;
 	}
 	
 	/** 
@@ -672,7 +703,7 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the mass of this body
 	 */
 	public float getMass() {
-		return  (body != null ? body.getMass() : super.getMass());
+		return massdata.mass;
 	}
 	
 	/** 
@@ -683,59 +714,72 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @param value  the mass of this body
 	 */
 	public void setMass(float value) {
-		super.setMass(value);
-		if (body != null) {
-			body.setMassData(massdata); // Protected accessor?
+		if (!masseffect) {
+			masseffect = true;
+			massdata.center.set(getCentroid());
+			massdata.I = getInertia();
 		}
+		massdata.mass = value;
 	}
 	
 	/**
 	 * Resets this body to use the mass computed from the its shape and density
 	 */
 	public void resetMass() {
-		super.resetMass();
-		if (body != null) {
-			body.resetMassData();
-		}
+		masseffect = false;
 	}
 	
-	/// Texture Information
+	/// Garbage Collection Methods	
 	/**
-	 * Returns the object texture for drawing purposes.
+	 * Returns true if our object has been flagged for garbage collection
 	 *
-	 * In order for drawing to work properly, you MUST set the drawScale.
-	 * The drawScale converts the physics units to pixels.
-	 * 
-	 * @return the object texture for drawing purposes.
+	 * A garbage collected object will be removed from the physics world at
+	 * the next time step.
+	 *
+	 * @return true if our object has been flagged for garbage collection
 	 */
-	public TextureRegion getTexture() {
-		return texture;
+	public boolean isRemoved() {
+		return toRemove;
 	}
 	
 	/**
-	 * Sets the object texture for drawing purposes.
+	 * Sets whether our object has been flagged for garbage collection
 	 *
-	 * In order for drawing to work properly, you MUST set the drawScale.
-	 * The drawScale converts the physics units to pixels.
-	 * 
-	 * @param value  the object texture for drawing purposes.
+	 * A garbage collected object will be removed from the physics world at
+	 * the next time step.
+	 *
+	 * @param value  whether our object has been flagged for garbage collection
 	 */
-	public void setTexture(TextureRegion value) {
-		texture = value;
-		origin.set(texture.getRegionWidth()/2.0f, texture.getRegionHeight()/2.0f);
+	public void markRemoved(boolean value) {
+		toRemove = value;
 	}
 	
 	/**
-	 * Draws the physics object.
+	 * Returns true if the shape information must be updated.
 	 *
-	 * @param canvas Drawing context
+	 * Attributes tied to the geometry (and not just forces/position) must wait for
+	 * collisions to complete before they are reset.  Shapes (and their properties) 
+	 * are reset in the update method.
+	 *
+	 * @return true if the shape information must be updated.
 	 */
-	public void draw(GameCanvas canvas) {
-		if (texture != null) {
-			canvas.draw(texture,Color.WHITE,origin.x,origin.y,getX()*drawScale.x,getY()*drawScale.x,getAngle(),1,1);
-		}
+	public boolean isDirty() {
+		return isDirty;
 	}
 	
+	/**
+	 * Sets whether the shape information must be updated.
+	 *
+	 * Attributes tied to the geometry (and not just forces/position) must wait for
+	 * collisions to complete before they are reset.  Shapes (and their properties) 
+	 * are reset in the update method.
+	 *
+	 * @param value  whether the shape information must be updated.
+	 */
+	public void markDirty(boolean value) {
+		isDirty = value;
+	}
+		
 	/**
 	 * Returns the Box2D body for this object.
 	 *
@@ -744,30 +788,127 @@ public abstract class SimpleGameObject extends GameObject {
 	 * @return the Box2D body for this object.
 	 */
 	public Body getBody() {
-		return body;
+		return null;
+	}
+	
+	/// DRAWING METHODS
+	/**
+     * Returns the drawing scale for this physics object
+     *
+     * The drawing scale is the number of pixels to draw before Box2D unit. Because
+     * mass is a function of area in Box2D, we typically want the physics objects
+     * to be small.  So we decouple that scale from the physics object.  However,
+     * we must track the scale difference to communicate with the scene graph.
+     *
+	 * This method does NOT return a reference to the drawing scale. Changes to this 
+	 * vector will not affect the body.  However, it returns the same vector each time
+	 * its is called, and so cannot be used as an allocator.
+
+     * We allow for the scaling factor to be non-uniform.
+     *
+     * @return the drawing scale for this physics object
+     */
+    public Vector2 getDrawScale() { 
+    	scaleCache.set(drawScale);
+    	return scaleCache; 
+    }
+    
+    /**
+     * Sets the drawing scale for this physics object
+     *
+     * The drawing scale is the number of pixels to draw before Box2D unit. Because
+     * mass is a function of area in Box2D, we typically want the physics objects
+     * to be small.  So we decouple that scale from the physics object.  However,
+     * we must track the scale difference to communicate with the scene graph.
+     *
+     * We allow for the scaling factor to be non-uniform.
+     *
+     * @param value  the drawing scale for this physics object
+     */
+    public void setDrawScale(Vector2 value) { 
+    	setDrawScale(value.x,value.y); 
+	}
+    
+    /**
+     * Sets the drawing scale for this physics object
+     *
+     * The drawing scale is the number of pixels to draw before Box2D unit. Because
+     * mass is a function of area in Box2D, we typically want the physics objects
+     * to be small.  So we decouple that scale from the physics object.  However,
+     * we must track the scale difference to communicate with the scene graph.
+     *
+     * We allow for the scaling factor to be non-uniform.
+     *
+     * @param x  the x-axis scale for this physics object
+     * @param y  the y-axis scale for this physics object
+     */
+    public void setDrawScale(float x, float y) {
+    	drawScale.set(x,y);
+    }
+    	
+	/// DEBUG METHODS
+	/**
+     * Returns the physics object tag.
+     * 
+     * A tag is a string attached to an object, in order to identify it in debugging.
+     *
+     * @return the physics object tag.
+     */
+    public String getName() { 
+    	return nametag; 
+    }
+    
+    /**
+     * Sets the physics object tag.
+     *
+     * A tag is a string attached to an object, in order to identify it in debugging.
+     *
+     * @param  value    the physics object tag
+     */
+    public void setName(String value) {
+     	nametag = value; 
 	}
 	
 	/**
-	 * Creates a new simple physics object at the origin.
-	 * 
-	 * REMEMBER: The size is in physics units, not pixels.
+	 * Create a new physics object at the origin.
 	 */
-	protected SimpleGameObject() {
+	protected Obstacle() { 
 		this(0,0);
 	}
-	
+
 	/**
-	 * Creates a new simple physics object
+	 * Create a new physics object
 	 * 
-	 * @param x  Initial x position in world coordinates
-	 * @param y  Initial y position in world coordinates
+	 * @param x Initial x position in world coordinates
+	 * @param y Initial y position in world coordinates
 	 */
-	protected SimpleGameObject(float x, float y) {
-		super(x,y);
-		origin = new Vector2();
-		body = null;
+	protected Obstacle(float x, float y) {
+		// Object has yet to be deactivated
+		toRemove = false;
+		
+		// Allocate the body information
+		bodyinfo = new BodyDef();
+		bodyinfo.awake  = true;
+		bodyinfo.allowSleep = true;
+		bodyinfo.gravityScale = 1.0f;
+		bodyinfo.position.set(x,y);
+		bodyinfo.fixedRotation = false;
+		// Objects are physics objects unless otherwise noted
+		bodyinfo.type = BodyType.DynamicBody;
+		
+		// Allocate the fixture information
+		// Default values are okay
+		fixture = new FixtureDef();
+
+		// Allocate the mass information, but turn it off
+		masseffect = false;
+		massdata = new MassData();
+		
+		// Set the default drawing scale
+		drawScale = new Vector2(1,1);
 	}
-	
+
+	/// Abstract Methods
 	/**
 	 * Creates the physics Body(s) for this object, adding them to the world.
 	 *
@@ -778,53 +919,16 @@ public abstract class SimpleGameObject extends GameObject {
 	 *
 	 * @return true if object allocation succeeded
 	 */
-	public boolean activatePhysics(World world) {
-		// Make a body, if possible
-		bodyinfo.active = true;
-		body = world.createBody(bodyinfo);
-		body.setUserData(this);
-		
-		// Only initialize if a body was created.
-		if (body != null) {
-			createFixtures();
-			return true;
-		} 
-		
-		bodyinfo.active = false;
-		return false;
-	}
-	
+	public abstract boolean activatePhysics(World world);
+
 	/**
 	 * Destroys the physics Body(s) of this object if applicable,
 	 * removing them from the world.
 	 * 
 	 * @param world Box2D world that stores body
 	 */
-	public void deactivatePhysics(World world) {
-		// Should be good for most (simple) applications.
-		if (body != null) {
-			// Snapshot the values
-			setBodyState(body);
-			world.destroyBody(body);
-			body = null;
-			bodyinfo.active = false;
-		}
-	}
+	public abstract void deactivatePhysics(World world);
 
-	/**
-	 * Create new fixtures for this body, defining the shape
-	 *
-	 * This is the primary method to override for custom physics objects
-	 */
-	protected abstract void createFixtures();
-
-    /**
-     * Release the fixtures for this body, reseting the shape
-     *
-     * This is the primary method to override for custom physics objects.
-     */
-	protected abstract void releaseFixtures();
-    
 	/**
 	 * Updates the object's physics state (NOT GAME LOGIC).
 	 *
@@ -835,10 +939,23 @@ public abstract class SimpleGameObject extends GameObject {
 	 *
 	 * @param dt Timing values from parent loop
 	 */
-	public void update(float delta) {
-		// Recreate the fixture object if dimensions changed.
-		if (isDirty()) {
-			createFixtures();
-		}
+	public void update(float delta) { 
 	}
+
+	/**
+	 * Draws the texture physics object.
+	 *
+	 * @param canvas Drawing context
+	 */
+	public abstract void draw(GameCanvas canvas);
+
+	/**
+	 * Draws the outline of the physics body.
+	 *
+	 * This method can be helpful for understanding issues with collisions.
+	 *
+	 * @param canvas Drawing context
+	 */
+	public abstract void drawDebug(GameCanvas canvas);
+
 }
