+++ core/src/edu/cornell/gdiac/rabbeat/obstacle/SimpleObstacle.java
@@ -1,12 +1,11 @@
 /*
- * Obstacle.java
+ * SimpleObstacle.java
  *
- * This class and all of the others in this package are provided to give you easy 
- * access to Box2d.  Box2d's separation of bodies and shapes (e.g. fixtures) can 
- * be a bit daunting. This class combines them together to simplify matters.
+ * This class is a subclass of Obstacle that supports only one Body.
+ * it is the prime subclass of most models in the game.
  *
- * This, and its superclasses, are fairly robust.  You may want to use
- * them in your game.  They make Box2D a lot easier.
+ * This class does not provide Shape information, and cannot be instantiated
+ * directly.
  *
  * Many of the method comments in this class are taken from the Box2d manual by 
  * Erin Catto (2011).
@@ -15,59 +14,34 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * LibGDX version, 2/6/2015
  */
-package edu.cornell.gdiac.rabbeat.obstacles;
+package edu.cornell.gdiac.rabbeat.obstacle;
 
 import com.badlogic.gdx.math.*;
+import com.badlogic.gdx.graphics.Color;
+import com.badlogic.gdx.graphics.g2d.*;
 import com.badlogic.gdx.physics.box2d.*;
 import com.badlogic.gdx.physics.box2d.BodyDef.BodyType;
 
-import edu.cornell.gdiac.rabbeat.*;  // For GameCanvas
+import edu.cornell.gdiac.rabbeat.GameCanvas;
 
 /**
  * Base model class to support collisions.
  *
- * Instances represents a body and/or a group of bodies.
- * There should be NO game controlling logic code in a physics objects, 
- * that should reside in the Controllers.
+ * This is an instance of a Obstacle with just one body. It does not have any joints.
+ * It is the primary type of physics object. 
  *
- * This abstract class has no Body or Shape information and should never 
- * be instantiated directly. Instead, you should instantiate either
- * SimplePhysicsObject or ComplexPhysicsObject.  This class only exists
- * to unify common functionality. In particular, it wraps the body and 
- * and fixture information into a single interface.
+ * This class does not provide Shape information, and cannot be instantiated directly.
  */
-public abstract class GameObject {
-	/// Initialization structures to store body information
-	/** Stores the body information for this shape */
-	protected BodyDef bodyinfo;
-	/** Stores the fixture information for this shape */
-	protected FixtureDef fixture;
-	/** The mass data of this shape (which may override the fixture) */
-	protected MassData massdata;
-	/** Whether or not to use the custom mass data */
-	protected boolean masseffect;
-    /** A tag for debugging purposes */
-    private String nametag;
-	/** Drawing scale to convert physics units to pixels */
-	protected Vector2 drawScale;
+public abstract class SimpleObstacle extends Obstacle {
+	/** The physics body for Box2D. */
+	protected Body body;
 
-	/// Track garbage collection status
-	/** Whether the object should be removed from the world on next pass */
-	private boolean toRemove;
-	/** Whether the object has changed shape and needs a new fixture */
-	private boolean isDirty;
+	/** The texture for the shape. */
+	protected TextureRegion texture;
+
+	/** The texture origin for drawing */
+	protected Vector2 origin;
 	
-	/// Caching objects
-	/** A cache value for when the user wants to access the body position */
-	protected Vector2 positionCache = new Vector2();
-	/** A cache value for when the user wants to access the linear velocity */
-	protected Vector2 velocityCache = new Vector2();
-	/** A cache value for when the user wants to access the center of mass */
-	protected Vector2 centroidCache = new Vector2();
-	/** A cache value for when the user wants to access the drawing scale */
-	protected Vector2 scaleCache = new Vector2();
-
-
 	/// BodyDef Methods
 	/**
 	 * Returns the body type for Box2D physics
@@ -79,7 +53,7 @@ public abstract class GameObject {
 	 * @return the body type for Box2D physics
 	 */
 	public BodyType getBodyType() {
-		return bodyinfo.type;
+		return (body != null ? body.getType() : super.getBodyType());
 	}
 	
 	/**
@@ -92,7 +66,11 @@ public abstract class GameObject {
 	 * @return the body type for Box2D physics
 	 */
 	public void setBodyType(BodyType value) {
-		bodyinfo.type = value;
+		if (body != null) {
+			body.setType(value);
+		} else {
+			super.setBodyType(value);
+		}
 	}
 	
 	/**
@@ -105,7 +83,7 @@ public abstract class GameObject {
 	 * @return the current position for this physics body
 	 */
 	public Vector2 getPosition() {
-		return positionCache.set(bodyinfo.position);
+		return (body != null ? body.getPosition() : super.getPosition());
 	}
 	
 	/**
@@ -116,7 +94,11 @@ public abstract class GameObject {
 	 * @param value  the current position for this physics body
 	 */
 	public void setPosition(Vector2 value) {
-		bodyinfo.position.set(value);
+		if (body != null) {
+			body.setTransform(value,body.getAngle());
+		} else {
+			super.setPosition(value);
+		}
 	}
 
 	/**
@@ -126,7 +108,12 @@ public abstract class GameObject {
 	 * @param y  the y-coordinate for this physics body
 	 */
 	public void setPosition(float x, float y) {
-		bodyinfo.position.set(x, y);
+		if (body != null) {
+			positionCache.set(x,y);
+			body.setTransform(positionCache,body.getAngle());
+		} else {
+			super.setPosition(x,y);
+		}
 	}
 
 	/**
@@ -135,7 +122,7 @@ public abstract class GameObject {
 	 * @return the x-coordinate for this physics body
 	 */
 	public float getX() {
-		return bodyinfo.position.x;
+		return (body != null ? body.getPosition().x : super.getX());
 	}
 	
 	/**
@@ -144,7 +131,12 @@ public abstract class GameObject {
 	 * @param value  the x-coordinate for this physics body
 	 */
 	public void setX(float value) {
-		bodyinfo.position.x = value;
+		if (body != null) {
+			positionCache.set(value,body.getPosition().y);
+			body.setTransform(positionCache,body.getAngle());
+		} else {
+			super.setX(value);
+		}
 	}
 	
 	/**
@@ -153,7 +145,7 @@ public abstract class GameObject {
 	 * @return the y-coordinate for this physics body
 	 */
 	public float getY() {
-		return bodyinfo.position.y;
+		return (body != null ? body.getPosition().y : super.getY());
 	}
 	
 	/**
@@ -162,7 +154,12 @@ public abstract class GameObject {
 	 * @param value  the y-coordinate for this physics body
 	 */
 	public void setY(float value) {
-		bodyinfo.position.y = value;
+		if (body != null) {
+			positionCache.set(body.getPosition().x,value);
+			body.setTransform(positionCache,body.getAngle());
+		} else {
+			super.setY(value);
+		}
 	}
 	
 	/**
@@ -173,7 +170,7 @@ public abstract class GameObject {
 	 * @return the angle of rotation for this body
 	 */
 	public float getAngle() {
-		return bodyinfo.angle;
+		return (body != null ? body.getAngle() : super.getAngle());
 	}
 	
 	/**
@@ -182,7 +179,11 @@ public abstract class GameObject {
 	 * @param value  the angle of rotation for this body (in radians)
 	 */
 	public void setAngle(float value) {
-		bodyinfo.angle = value;
+		if (body != null) {
+			body.setTransform(body.getPosition(),value);
+		} else {
+			super.setAngle(value);
+		}
 	}
 	
 	/**
@@ -195,7 +196,7 @@ public abstract class GameObject {
 	 * @return the linear velocity for this physics body
 	 */
 	public Vector2 getLinearVelocity() {
-		return velocityCache.set(bodyinfo.linearVelocity);
+		return (body != null ? body.getLinearVelocity() : super.getLinearVelocity());
 	}
 	
 	/**
@@ -206,7 +207,11 @@ public abstract class GameObject {
 	 * @param value  the linear velocity for this physics body
 	 */
 	public void setLinearVelocity(Vector2 value) {
-		bodyinfo.linearVelocity.set(value);
+		if (body != null) {
+			body.setLinearVelocity(value);
+		} else {
+			super.setLinearVelocity(value);
+		}
 	}
 	
 	/**
@@ -215,7 +220,7 @@ public abstract class GameObject {
 	 * @return the x-velocity for this physics body
 	 */
 	public float getVX() {
-		return bodyinfo.linearVelocity.x;
+		return (body != null ? body.getLinearVelocity().x : super.getVX());
 	}
 	
 	/**
@@ -224,7 +229,12 @@ public abstract class GameObject {
 	 * @param value  the x-velocity for this physics body
 	 */
 	public void setVX(float value) {
-		bodyinfo.linearVelocity.x = value;
+		if (body != null) {
+			velocityCache.set(value,body.getLinearVelocity().y);
+			body.setLinearVelocity(velocityCache);
+		} else {
+			super.setVX(value);
+		}
 	}
 	
 	/**
@@ -233,7 +243,7 @@ public abstract class GameObject {
 	 * @return the y-velocity for this physics body
 	 */
 	public float getVY() {
-		return bodyinfo.linearVelocity.y;
+		return (body != null ? body.getLinearVelocity().y : super.getVY());
 	}
 	
 	/**
@@ -242,7 +252,12 @@ public abstract class GameObject {
 	 * @param value  the y-velocity for this physics body
 	 */
 	public void setVY(float value) {
-		bodyinfo.linearVelocity.y = value;
+		if (body != null) {
+			velocityCache.set(body.getLinearVelocity().x,value);
+			body.setLinearVelocity(velocityCache);
+		} else {
+			super.setVY(value);
+		}
 	}
 	
 	/**
@@ -253,7 +268,7 @@ public abstract class GameObject {
 	 * @return the angular velocity for this physics body
 	 */
 	public float getAngularVelocity() {
-		return bodyinfo.angularVelocity;
+		return (body != null ? body.getAngularVelocity() : super.getAngularVelocity());
 	}
 	
 	/**
@@ -262,7 +277,11 @@ public abstract class GameObject {
 	 * @param value the angular velocity for this physics body (in radians)
 	 */
 	public void setAngularVelocity(float value) {
-		bodyinfo.angularVelocity = value;
+		if (body != null) {
+			body.setAngularVelocity(value);
+		} else {
+			super.setAngularVelocity(value);
+		}
 	}
 	
 	/**
@@ -276,7 +295,7 @@ public abstract class GameObject {
 	 * @return true if the body is active
 	 */
 	public boolean isActive() {
-		return bodyinfo.active;
+		return (body != null ? body.isActive() : super.isActive());
 	}
 	
 	/**
@@ -290,7 +309,11 @@ public abstract class GameObject {
 	 * @param value  whether the body is active
 	 */
 	public void setActive(boolean value) {
-		bodyinfo.active = value;
+		if (body != null) {
+			body.setActive(value);
+		} else {
+			super.setActive(value);
+		}
 	}
 	
 	/**
@@ -304,7 +327,7 @@ public abstract class GameObject {
 	 * @return true if the body is awake
 	 */
 	public boolean isAwake() {
-		return bodyinfo.awake;
+		return (body != null ? body.isAwake() : super.isAwake());
 	}
 	
 	/**
@@ -318,7 +341,11 @@ public abstract class GameObject {
 	 * @param value  whether the body is awake
 	 */
 	public void setAwake(boolean value) {
-		bodyinfo.awake = value;
+		if (body != null) {
+			body.setAwake(value);
+		} else {
+			super.setAwake(value);
+		}
 	}
 	
 	/**
@@ -332,7 +359,7 @@ public abstract class GameObject {
 	 * @return false if this body should never fall asleep
 	 */
 	public boolean isSleepingAllowed() {
-		return bodyinfo.allowSleep;
+		return (body != null ? body.isSleepingAllowed() : super.isSleepingAllowed());
 	}
 	
 	/**
@@ -346,7 +373,11 @@ public abstract class GameObject {
 	 * @param value  whether the body should ever fall asleep
 	 */
 	public void setSleepingAllowed(boolean value) {
-		bodyinfo.allowSleep = value;
+		if (body != null) {
+			body.setSleepingAllowed(value);
+		} else {
+			super.setSleepingAllowed(value);
+		}
 	}
 	
 	/**
@@ -357,7 +388,7 @@ public abstract class GameObject {
 	 * @return true if this body be prevented from rotating
 	 */
 	public boolean isFixedRotation() {
-		return bodyinfo.fixedRotation;
+		return (body != null ? body.isFixedRotation() : super.isFixedRotation());
 	}
 	
 	/**
@@ -368,7 +399,11 @@ public abstract class GameObject {
 	 * @param value  whether this body be prevented from rotating
 	 */
 	public void setFixedRotation(boolean value) {
-		bodyinfo.fixedRotation = value;
+		if (body != null) {
+			body.setFixedRotation(value);
+		} else {
+			super.setFixedRotation(value);
+		}
 	}
 	
 	/**
@@ -380,7 +415,7 @@ public abstract class GameObject {
 	 * @return the gravity scale to apply to this body
 	 */
 	public float getGravityScale() {
-		return bodyinfo.gravityScale;
+		return (body != null ? body.getGravityScale() : super.getGravityScale());
 	}
 	
 	/**
@@ -392,7 +427,11 @@ public abstract class GameObject {
 	 * @param value  the gravity scale to apply to this body
 	 */
 	public void setGravityScale(float value) {
-		bodyinfo.gravityScale = value;
+		if (body != null) {
+			body.setGravityScale(value);
+		} else {
+			super.setGravityScale(value);
+		}
 	}
 	
 	/** 
@@ -409,7 +448,7 @@ public abstract class GameObject {
 	 * @return the linear damping for this body.
 	 */
 	public float getLinearDamping() {
-		return bodyinfo.linearDamping;
+		return (body != null ? body.getLinearDamping() : super.getLinearDamping());
 	}
 	
 	/** 
@@ -426,7 +465,11 @@ public abstract class GameObject {
 	 * @param value  the linear damping for this body.
 	 */
 	public void setLinearDamping(float value) {
-		bodyinfo.linearDamping = value;
+		if (body != null) {
+			body.setLinearDamping(value);
+		} else {
+			super.setLinearDamping(value);
+		}
 	}
 	
 	/** 
@@ -443,7 +486,7 @@ public abstract class GameObject {
 	 * @return the angular damping for this body.
 	 */
 	public float getAngularDamping() {
-		return bodyinfo.angularDamping;
+		return (body != null ? body.getAngularDamping() : super.getAngularDamping());
 	}
 
 	/** 
@@ -460,43 +503,14 @@ public abstract class GameObject {
 	 * @param value  the angular damping for this body.
 	 */
 	public void setAngularDamping(float value) {
-		bodyinfo.angularDamping = value;
-	}
-	
-	/**
-	 * Copies the state from the given body to the body def.
-	 *
-	 * This is important if you want to save the state of the body before removing
-	 * it from the world.
-	 */
-	protected void setBodyState(Body body) {
-		bodyinfo.type   = body.getType();
-		bodyinfo.angle  = body.getAngle();
-		bodyinfo.active = body.isActive();
-		bodyinfo.awake  = body.isAwake();
-		bodyinfo.position.set(body.getPosition());
-		bodyinfo.linearVelocity.set(body.getLinearVelocity());
-		bodyinfo.allowSleep = body.isSleepingAllowed();
-		bodyinfo.fixedRotation = body.isFixedRotation();
-		bodyinfo.gravityScale  = body.getGravityScale();
-		bodyinfo.angularDamping = body.getAngularDamping();
-		bodyinfo.linearDamping  = body.getLinearDamping();
+		if (body != null) {
+			body.setAngularDamping(value);
+		} else {
+			super.setAngularDamping(value);
+		}
 	}
 	
 	/// FixtureDef Methods
-	/**
-	 * Returns the density of this body
-	 *
-	 * The density is typically measured in usually in kg/m^2. The density can be zero or 
-	 * positive. You should generally use similar densities for all your fixtures. This 
-	 * will improve stacking stability.
-	 *
-	 * @return the density of this body
-	 */
-	public float getDensity() {
-		return fixture.density;
-	}
-	
 	/**
 	 * Sets the density of this body
 	 *
@@ -507,22 +521,12 @@ public abstract class GameObject {
 	 * @param value  the density of this body
 	 */
 	public void setDensity(float value) {
-		fixture.density = value;
-	}
-	
-	/**
-	 * Returns the friction coefficient of this body
-	 *
-	 * The friction parameter is usually set between 0 and 1, but can be any non-negative 
-	 * value. A friction value of 0 turns off friction and a value of 1 makes the friction 
-	 * strong. When the friction force is computed between two shapes, Box2D must combine 
-	 * the friction parameters of the two parent fixtures. This is done with the geometric 
-	 * mean.
-	 *
-	 * @return the friction coefficient of this body
-	 */
-	public float getFriction() {
-		return fixture.friction; 
+		super.setDensity(value);
+		if (body != null) {
+			for(Fixture f : body.getFixtureList()) {
+				f.setDensity(value);
+			}
+		}
 	}
 	
 	/**
@@ -537,22 +541,12 @@ public abstract class GameObject {
 	 * @param value  the friction coefficient of this body
 	 */
 	public void setFriction(float value) {
-		fixture.friction = value;
-	}
-	
-	/**
-	 * Returns the restitution of this body
-	 *
-	 * Restitution is used to make objects bounce. The restitution value is usually set 
-	 * to be between 0 and 1. Consider dropping a ball on a table. A value of zero means 
-	 * the ball won't bounce. This is called an inelastic collision. A value of one means 
-	 * the ball's velocity will be exactly reflected. This is called a perfectly elastic 
-	 * collision.
-	 *
-	 * @return the restitution of this body
-	 */
-	public float getRestitution() {
-		return fixture.restitution;
+		super.setFriction(value);
+		if (body != null) {
+			for(Fixture f : body.getFixtureList()) {
+				f.setFriction(value);
+			}
+		}
 	}
 	
 	/**
@@ -567,20 +561,12 @@ public abstract class GameObject {
 	 * @param value  the restitution of this body
 	 */
 	public void setRestitution(float value) {
-		fixture.restitution = value;
-	}
-	
-	/**
-	 * Returns true if this object is a sensor.
-	 *
-	 * Sometimes game logic needs to know when two entities overlap yet there should be 
-	 * no collision response. This is done by using sensors. A sensor is an entity that 
-	 * detects collision but does not produce a response.
-	 *
-	 * @return true if this object is a sensor.
-	 */
-	public boolean isSensor() {
-		return fixture.isSensor;
+		super.setRestitution(value);
+		if (body != null) {
+			for(Fixture f : body.getFixtureList()) {
+				f.setRestitution(value);
+			}
+		}
 	}
 	
 	/**
@@ -593,22 +579,12 @@ public abstract class GameObject {
 	 * @param value  whether this object is a sensor.
 	 */
 	public void setSensor(boolean value) {
-		fixture.isSensor = value;
-	}
-	
-	/**
-	 * Returns the filter data for this object (or null if there is none)
-	 *
-	 * Collision filtering allows you to prevent collision between fixtures. For example, 
-	 * say you make a character that rides a bicycle. You want the bicycle to collide 
-	 * with the terrain and the character to collide with the terrain, but you don't want 
-	 * the character to collide with the bicycle (because they must overlap). Box2D 
-	 * supports such collision filtering using categories and groups.
-	 * 
-	 * @return the filter data for this object (or null if there is none)
-	 */
-	public Filter getFilterData() {
-		return fixture.filter;
+		super.setSensor(value);
+		if (body != null) {
+			for(Fixture f : body.getFixtureList()) {
+				f.setSensor(value);
+			}
+		}
 	}
 	
 	/**
@@ -625,14 +601,11 @@ public abstract class GameObject {
 	 * @param value  the filter data for this object
 	 */
 	public void setFilterData(Filter value) {
-		if (value !=  null) {
-			fixture.filter.categoryBits = value.categoryBits;
-			fixture.filter.groupIndex = value.groupIndex;
-			fixture.filter.maskBits   = value.maskBits;
-		} else {
-			fixture.filter.categoryBits = 0x0001;
-			fixture.filter.groupIndex = 0;
-			fixture.filter.maskBits   = -1;
+		super.setFilterData(value);
+		if (body != null) {
+			for(Fixture f : body.getFixtureList()) {
+				f.setFilterData(value);
+			}
 		}
 	}
 
@@ -647,7 +620,7 @@ public abstract class GameObject {
 	 * @return the center of mass for this physics body
 	 */
 	public Vector2 getCentroid() {
-		return centroidCache.set(massdata.center);
+		return  (body != null ? body.getLocalCenter() : super.getCentroid());
 	}
 	
 	/**
@@ -658,12 +631,10 @@ public abstract class GameObject {
 	 * @param value  the center of mass for this physics body
 	 */
 	public void setCentroid(Vector2 value) {
-		if (!masseffect) {
-			masseffect = true;
-			massdata.I = getInertia();
-			massdata.mass = getMass();
+		super.setCentroid(value);
+		if (body != null) {
+			body.setMassData(massdata); // Protected accessor?
 		}
-		massdata.center.set(value);
 	}
 	
 	/**
@@ -675,7 +646,7 @@ public abstract class GameObject {
 	 * @return the rotational inertia of this body
 	 */
 	public float getInertia() {
-		return massdata.I;
+		return  (body != null ? body.getInertia() : super.getInertia());
 	}
 
 	/**
@@ -687,12 +658,10 @@ public abstract class GameObject {
 	 * @param value  the rotational inertia of this body
 	 */
 	public void setInertia(float value) {
-		if (!masseffect) {
-			masseffect = true;
-			massdata.center.set(getCentroid());
-			massdata.mass = getMass();
+		super.setInertia(value);
+		if (body != null) {
+			body.setMassData(massdata); // Protected accessor?
 		}
-		massdata.I = value;
 	}
 	
 	/** 
@@ -703,7 +672,7 @@ public abstract class GameObject {
 	 * @return the mass of this body
 	 */
 	public float getMass() {
-		return massdata.mass;
+		return  (body != null ? body.getMass() : super.getMass());
 	}
 	
 	/** 
@@ -714,72 +683,59 @@ public abstract class GameObject {
 	 * @param value  the mass of this body
 	 */
 	public void setMass(float value) {
-		if (!masseffect) {
-			masseffect = true;
-			massdata.center.set(getCentroid());
-			massdata.I = getInertia();
+		super.setMass(value);
+		if (body != null) {
+			body.setMassData(massdata); // Protected accessor?
 		}
-		massdata.mass = value;
 	}
 	
 	/**
 	 * Resets this body to use the mass computed from the its shape and density
 	 */
 	public void resetMass() {
-		masseffect = false;
+		super.resetMass();
+		if (body != null) {
+			body.resetMassData();
+		}
 	}
 	
-	/// Garbage Collection Methods	
+	/// Texture Information
 	/**
-	 * Returns true if our object has been flagged for garbage collection
+	 * Returns the object texture for drawing purposes.
 	 *
-	 * A garbage collected object will be removed from the physics world at
-	 * the next time step.
-	 *
-	 * @return true if our object has been flagged for garbage collection
+	 * In order for drawing to work properly, you MUST set the drawScale.
+	 * The drawScale converts the physics units to pixels.
+	 * 
+	 * @return the object texture for drawing purposes.
 	 */
-	public boolean isRemoved() {
-		return toRemove;
-	}
-	
-	/**
-	 * Sets whether our object has been flagged for garbage collection
-	 *
-	 * A garbage collected object will be removed from the physics world at
-	 * the next time step.
-	 *
-	 * @param value  whether our object has been flagged for garbage collection
-	 */
-	public void markRemoved(boolean value) {
-		toRemove = value;
+	public TextureRegion getTexture() {
+		return texture;
 	}
 	
 	/**
-	 * Returns true if the shape information must be updated.
+	 * Sets the object texture for drawing purposes.
 	 *
-	 * Attributes tied to the geometry (and not just forces/position) must wait for
-	 * collisions to complete before they are reset.  Shapes (and their properties) 
-	 * are reset in the update method.
-	 *
-	 * @return true if the shape information must be updated.
+	 * In order for drawing to work properly, you MUST set the drawScale.
+	 * The drawScale converts the physics units to pixels.
+	 * 
+	 * @param value  the object texture for drawing purposes.
 	 */
-	public boolean isDirty() {
-		return isDirty;
+	public void setTexture(TextureRegion value) {
+		texture = value;
+		origin.set(texture.getRegionWidth()/2.0f, texture.getRegionHeight()/2.0f);
 	}
 	
 	/**
-	 * Sets whether the shape information must be updated.
+	 * Draws the physics object.
 	 *
-	 * Attributes tied to the geometry (and not just forces/position) must wait for
-	 * collisions to complete before they are reset.  Shapes (and their properties) 
-	 * are reset in the update method.
-	 *
-	 * @param value  whether the shape information must be updated.
+	 * @param canvas Drawing context
 	 */
-	public void markDirty(boolean value) {
-		isDirty = value;
+	public void draw(GameCanvas canvas) {
+		if (texture != null) {
+			canvas.draw(texture,Color.WHITE,origin.x,origin.y,getX()*drawScale.x,getY()*drawScale.x,getAngle(),1,1);
+		}
 	}
-		
+	
 	/**
 	 * Returns the Box2D body for this object.
 	 *
@@ -788,127 +744,30 @@ public abstract class GameObject {
 	 * @return the Box2D body for this object.
 	 */
 	public Body getBody() {
-		return null;
-	}
-	
-	/// DRAWING METHODS
-	/**
-     * Returns the drawing scale for this physics object
-     *
-     * The drawing scale is the number of pixels to draw before Box2D unit. Because
-     * mass is a function of area in Box2D, we typically want the physics objects
-     * to be small.  So we decouple that scale from the physics object.  However,
-     * we must track the scale difference to communicate with the scene graph.
-     *
-	 * This method does NOT return a reference to the drawing scale. Changes to this 
-	 * vector will not affect the body.  However, it returns the same vector each time
-	 * its is called, and so cannot be used as an allocator.
-
-     * We allow for the scaling factor to be non-uniform.
-     *
-     * @return the drawing scale for this physics object
-     */
-    public Vector2 getDrawScale() { 
-    	scaleCache.set(drawScale);
-    	return scaleCache; 
-    }
-    
-    /**
-     * Sets the drawing scale for this physics object
-     *
-     * The drawing scale is the number of pixels to draw before Box2D unit. Because
-     * mass is a function of area in Box2D, we typically want the physics objects
-     * to be small.  So we decouple that scale from the physics object.  However,
-     * we must track the scale difference to communicate with the scene graph.
-     *
-     * We allow for the scaling factor to be non-uniform.
-     *
-     * @param value  the drawing scale for this physics object
-     */
-    public void setDrawScale(Vector2 value) { 
-    	setDrawScale(value.x,value.y); 
-	}
-    
-    /**
-     * Sets the drawing scale for this physics object
-     *
-     * The drawing scale is the number of pixels to draw before Box2D unit. Because
-     * mass is a function of area in Box2D, we typically want the physics objects
-     * to be small.  So we decouple that scale from the physics object.  However,
-     * we must track the scale difference to communicate with the scene graph.
-     *
-     * We allow for the scaling factor to be non-uniform.
-     *
-     * @param x  the x-axis scale for this physics object
-     * @param y  the y-axis scale for this physics object
-     */
-    public void setDrawScale(float x, float y) {
-    	drawScale.set(x,y);
-    }
-    	
-	/// DEBUG METHODS
-	/**
-     * Returns the physics object tag.
-     * 
-     * A tag is a string attached to an object, in order to identify it in debugging.
-     *
-     * @return the physics object tag.
-     */
-    public String getName() { 
-    	return nametag; 
-    }
-    
-    /**
-     * Sets the physics object tag.
-     *
-     * A tag is a string attached to an object, in order to identify it in debugging.
-     *
-     * @param  value    the physics object tag
-     */
-    public void setName(String value) {
-     	nametag = value; 
+		return body;
 	}
 	
 	/**
-	 * Create a new physics object at the origin.
+	 * Creates a new simple physics object at the origin.
+	 * 
+	 * REMEMBER: The size is in physics units, not pixels.
 	 */
-	protected GameObject() {
+	protected SimpleObstacle() {
 		this(0,0);
 	}
-
+	
 	/**
-	 * Create a new physics object
+	 * Creates a new simple physics object
 	 * 
-	 * @param x Initial x position in world coordinates
-	 * @param y Initial y position in world coordinates
+	 * @param x  Initial x position in world coordinates
+	 * @param y  Initial y position in world coordinates
 	 */
-	protected GameObject(float x, float y) {
-		// Object has yet to be deactivated
-		toRemove = false;
-		
-		// Allocate the body information
-		bodyinfo = new BodyDef();
-		bodyinfo.awake  = true;
-		bodyinfo.allowSleep = true;
-		bodyinfo.gravityScale = 1.0f;
-		bodyinfo.position.set(x,y);
-		bodyinfo.fixedRotation = false;
-		// Objects are physics objects unless otherwise noted
-		bodyinfo.type = BodyType.DynamicBody;
-		
-		// Allocate the fixture information
-		// Default values are okay
-		fixture = new FixtureDef();
-
-		// Allocate the mass information, but turn it off
-		masseffect = false;
-		massdata = new MassData();
-		
-		// Set the default drawing scale
-		drawScale = new Vector2(1,1);
+	protected SimpleObstacle(float x, float y) {
+		super(x,y);
+		origin = new Vector2();
+		body = null;
 	}
-
-	/// Abstract Methods
+	
 	/**
 	 * Creates the physics Body(s) for this object, adding them to the world.
 	 *
@@ -919,16 +778,53 @@ public abstract class GameObject {
 	 *
 	 * @return true if object allocation succeeded
 	 */
-	public abstract boolean activatePhysics(World world);
-
+	public boolean activatePhysics(World world) {
+		// Make a body, if possible
+		bodyinfo.active = true;
+		body = world.createBody(bodyinfo);
+		body.setUserData(this);
+		
+		// Only initialize if a body was created.
+		if (body != null) {
+			createFixtures();
+			return true;
+		} 
+		
+		bodyinfo.active = false;
+		return false;
+	}
+	
 	/**
 	 * Destroys the physics Body(s) of this object if applicable,
 	 * removing them from the world.
 	 * 
 	 * @param world Box2D world that stores body
 	 */
-	public abstract void deactivatePhysics(World world);
+	public void deactivatePhysics(World world) {
+		// Should be good for most (simple) applications.
+		if (body != null) {
+			// Snapshot the values
+			setBodyState(body);
+			world.destroyBody(body);
+			body = null;
+			bodyinfo.active = false;
+		}
+	}
 
+	/**
+	 * Create new fixtures for this body, defining the shape
+	 *
+	 * This is the primary method to override for custom physics objects
+	 */
+	protected abstract void createFixtures();
+
+    /**
+     * Release the fixtures for this body, reseting the shape
+     *
+     * This is the primary method to override for custom physics objects.
+     */
+	protected abstract void releaseFixtures();
+    
 	/**
 	 * Updates the object's physics state (NOT GAME LOGIC).
 	 *
@@ -939,23 +835,10 @@ public abstract class GameObject {
 	 *
 	 * @param dt Timing values from parent loop
 	 */
-	public void update(float delta) { 
+	public void update(float delta) {
+		// Recreate the fixture object if dimensions changed.
+		if (isDirty()) {
+			createFixtures();
+		}
 	}
-
-	/**
-	 * Draws the texture physics object.
-	 *
-	 * @param canvas Drawing context
-	 */
-	public abstract void draw(GameCanvas canvas);
-
-	/**
-	 * Draws the outline of the physics body.
-	 *
-	 * This method can be helpful for understanding issues with collisions.
-	 *
-	 * @param canvas Drawing context
-	 */
-	public abstract void drawDebug(GameCanvas canvas);
-
 }
