+++ core/src/edu/cornell/gdiac/rabbeat/obstacle/enemies/BearEnemy.java
@@ -1,12 +1,11 @@
-package edu.cornell.gdiac.rabbeat.obstacles.enemies;
+package edu.cornell.gdiac.rabbeat.obstacle.enemies;
 
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.JsonValue;
-import edu.cornell.gdiac.rabbeat.GameController;
 import edu.cornell.gdiac.rabbeat.Genre;
-import edu.cornell.gdiac.rabbeat.ObjectController;
+import edu.cornell.gdiac.rabbeat.WorldController;
 import edu.cornell.gdiac.rabbeat.sync.BulletSync;
 import edu.cornell.gdiac.rabbeat.sync.ISynced;
 
@@ -16,7 +15,7 @@ public class BearEnemy extends Enemy implements ISynced {
     private int beatCount = 0;
 
     /** The bullet the bear will be shooting */
-    private Bullet bullet;
+    private BulletSync bullet;
 
     /** Speed of the bullet when the game is in synth mode */
     private float synthSpeed;
@@ -27,8 +26,14 @@ public class BearEnemy extends Enemy implements ISynced {
     /** Current genre that the game is on */
     private Genre curGenre = Genre.SYNTH;
 
+    /** JsonValue for bullet */
+    private JsonValue bulletjv =  WorldController.getInstance().getBulletJV();
+
+    /** Texture for bullet */
+    private TextureRegion bullettr = WorldController.getInstance().getBulletTR();
+
     /** Scale of the world */
-    private Vector2 scale = GameController.getInstance().getScale();
+    private Vector2 scale = WorldController.getInstance().getScale();
 
     /**
      * Creates a new enemy avatar with the given physics data
@@ -40,11 +45,11 @@ public class BearEnemy extends Enemy implements ISynced {
      * @param faceRight
      */
     public BearEnemy(JsonValue data, float width, float height, float enemyScale,
-            boolean faceRight) {
+            boolean faceRight, float synthVX, float jazzVX) {
         super(data, width, height, enemyScale, faceRight);
         float dir = (faceRight ? 1 : -1);
-        synthSpeed = data.get("max_speed").getFloat("synth") * dir;
-        jazzSpeed = data.get("max_speed").getFloat("jazz") * dir;
+        synthSpeed = synthVX * dir;
+        jazzSpeed = jazzVX * dir;
     }
 
     /**
@@ -70,36 +75,35 @@ public class BearEnemy extends Enemy implements ISynced {
 
     /** Creates a bullet in front of the bear */
     public void makeBullet(){
-          //TODO: create a bullet using object controller default values.  instantiate the copy using gamecontroller
+        float offset = bulletjv.getFloat("offset",0);
+        offset *= (isFaceRight() ? 1 : -1);
+        float radius = bullettr.getRegionWidth()/(2.0f*scale.x);
+        bullet = new BulletSync(getX()+offset, getY(), radius, bulletjv.getFloat("synth speed", 0),
+                bulletjv.getFloat("jazz speed", 0), isFaceRight());
 
-//        float offset = ObjectController.bullet.getFloat("offset",0);
-//        offset *= (isFaceRight() ? 1 : -1);
-//        float radius = bullettr.getRegionWidth()/(2.0f*scale.x);
-//        bullet = new BulletSync(getX()+offset, getY(), radius, bulletjv.getFloat("synth speed", 0),
-//                bulletjv.getFloat("jazz speed", 0), isFaceRight());
-//
-//        bullet.setName("bullet");
-//        bullet.setDensity(bulletjv.getFloat("density", 0));
-//        bullet.setDrawScale(scale);
-//        bullet.setTexture(bullettr);
-//        bullet.setGravityScale(0);
+        bullet.setName("bullet");
+        bullet.setDensity(bulletjv.getFloat("density", 0));
+        bullet.setDrawScale(scale);
+        bullet.setTexture(bullettr);
+        bullet.setGravityScale(0);
 
         // Compute position and velocity
-//        float speed;
-//        if (curGenre == Genre.SYNTH){
-//            speed = bulletjv.getFloat("synth speed", 0);
-//        }
-//        else {
-//            speed = bulletjv.getFloat("jazz speed", 0);
-//        }
-//        speed *= (isFaceRight() ? 1 : -1);
-//        bullet.setVX(speed);
-//        GameController.getInstance().instantiateQueue(bullet);
+        float speed;
+        if (curGenre == Genre.SYNTH){
+            speed = bulletjv.getFloat("synth speed", 0);
+        }
+        else {
+            speed = bulletjv.getFloat("jazz speed", 0);
+        }
+        speed *= (isFaceRight() ? 1 : -1);
+        bullet.setVX(speed);
+        WorldController.getInstance().instantiate(bullet);
     }
     public float getBeat() {
         return 1;
     }
 
+    @Override
     public void beatAction() {
         beatCount++;
         if(beatCount >= 5){
@@ -111,9 +115,10 @@ public class BearEnemy extends Enemy implements ISynced {
         if (beatCount == 4){
         }
 
-        setFaceRight(GameController.getInstance().getPlayer().getPosition().x - getPosition().x > 0);
+        setFaceRight(WorldController.getInstance().getPlayer().getPosition().x - getPosition().x > 0);
     }
 
+    @Override
     public void genreUpdate(Genre genre) {
         changeSpeed(genre);
     }
